componentize media placeholder
